Corrected the invocation of the "extract_files" method of ADFSdisc in the "__main__" part of the module so that the parameters are in the order expected by that method.
Note that the "extract_files" method takes its parameters in the reverse
order to the "extract_old_files" and "extract_new_files" methods because it
allows the user to avoid explicitly specifying which files are to be
extracted. When no files are specified then all files on the disc are
extracted.