Found a problem with disc titles containing control characters (\x00 in particular) on disc images sent by Simon Brooke <simon@jasmine.org.uk>.
The problem was introduced by the old format disc catalogue reading method
"read_old_catalogue" which uses the directory title of the sector whose
parent is itself as the disc name. No control code stripping was done and
it turned out that the "safe" method needed extending to allow spaces for
this purpose since they appear to be valid characters in disc titles.